modularize wip
rename http server dir

move wkhttp into static library module
