+++ .circleci/config.yml
@@ -12,14 +12,15 @@ jobs:
           mkdir build
           cd build
           cmake ..
-          make install
+          make
     - run:
         name: GTests
-        command: /opt/mct/bin/runUnitTests
+        command: ./build/runUnitTests
     - run:
         name: Py Integration Tests
         command: |
-          cd src/tests/py_int_test
+          pushd src/tests/py_int_test
           apt-get -y install python3 python3-pip python3-pycurl
           pip3 install -r requirements.txt
-          /opt/mct/bin/marine_chart_tileserver & python3 int_test.py
+          popd
+          ./build/exampleServer & python3 src/tests/py_int_test/int_test.py
