+++ .gitignore
@@ -2,3 +2,6 @@ scratch/
 cmake-build-debug/
 .idea/
 *.iml
+build/
+*.pyc
+__pycache__/
