+++ src/wkhttp/LambdaRequestHandlerFactory.cpp
