+++ src/wkhttp/LambdaRequestHandlerFactory.h
