+++ src/wkhttp/StaticContentHandlerFactory.cpp
