+++ src/wkhttp/StaticContentHandlerFactory.h
